Develop

+++ src/Services/whatsapp/WhatsAppClient.ts
@@ -4,6 +4,7 @@ import {Server as SocketIOServer} from 'socket.io'
 import ChatBot from '../chatBot/ChatBot'
 import {DataSnapshot} from 'firebase-admin/lib/database'
 import * as Messages from '../chatBot/Messages'
+import {ASK_FOR_CANCEL} from '../chatBot/Messages'
 import {Store} from '../store/Store'
 import config from '../../../config';
 import {WpNotificationType} from '../../Interfaces/WpNotificationType'
@@ -12,21 +13,34 @@ import {exit} from 'process'
 import {EmitEvents} from './EmitEvents'
 import {LoadingType} from '../../Interfaces/LoadingType'
 import SettingsRepository from '../../Repositories/SettingsRepository'
-import SessionRepository from '../../Repositories/SessionRepository'
 import ServiceRepository from '../../Repositories/ServiceRepository'
 import Service from '../../Models/Service'
-import {ASK_FOR_CANCEL} from '../chatBot/Messages'
+import {WpClient} from "../../Interfaces/WpClient";
+import Session from "../../Models/Session";
+import SessionRepository from "../../Repositories/SessionRepository";
+import {ServiceInterface} from "../../Interfaces/ServiceInterface";
 
-export default class WhatsAppClient {
+export class WhatsAppClient {
   
   public client: Client
   private socket: SocketIOServer | null = null
-  static SESSION_PATH = 'storage/sessions'
+  static SESSION_PATH = 'storage/sessions/'
   private chatBot: ChatBot
   private store: Store = Store.getInstance()
+	private wpClient: WpClient
+	public deleting = false
+	public starting = false
+
+	constructor(client: WpClient) {
+		this.wpClient = client
+	}
+
 	initClient(): void {
     this.client = new Client({
-      authStrategy: new LocalAuth({dataPath: WhatsAppClient.SESSION_PATH}),
+      authStrategy: new LocalAuth({
+				clientId: this.wpClient.id,
+				dataPath: WhatsAppClient.SESSION_PATH
+			}),
 			qrMaxRetries: 2,
 			takeoverOnConflict: false,
       puppeteer: {
@@ -38,7 +52,7 @@ export default class WhatsAppClient {
           '--disable-setuid-sandbox',
           '--disable-dev-shm-usage',
           '--unhandled-rejections=strict',
-		  '--no-zygote'
+		  		'--no-zygote'
         ]
       }
     })
@@ -49,18 +63,21 @@ export default class WhatsAppClient {
     this.client.on(Events.AUTHENTICATION_FAILURE, this.onAuthFailure)
     this.client.on(Events.STATE_CHANGED, this.onStateChanged)
     this.client.on(Events.DISCONNECTED, this.onDisconnected)
-	this.client.on(Events.LOADING_SCREEN, this.onLoadingScreen)
-	this.client.on(Events.MESSAGE_RECEIVED, this.onMessageReceived)
+		this.client.on(Events.LOADING_SCREEN, this.onLoadingScreen)
+		this.client.on(Events.MESSAGE_RECEIVED, this.onMessageReceived)
 
 	this.init(false)
 	  .then(async () => {
-		  console.log('authenticated after init server')
-		  await SettingsRepository.enableWpNotifications(true).catch(e => console.log(e.message))
+		  console.log('authenticated after init server', this.wpClient.alias)
+			this.starting = false
+		  if (!this.deleting) await SettingsRepository.enableWpNotifications(this.wpClient.id, true)
+			.catch(e => console.log(e.message))
 	  })
 	  .catch(e => {
+			this.starting = false
 			console.log(e.message)
-		Sentry.captureException(e)
-		exit(1)
+		  Sentry.captureException(e)
+		  exit(1)
 	  })
   }
   
@@ -74,39 +91,38 @@ export default class WhatsAppClient {
   
   onReady = (): void => {
     this.chatBot = new ChatBot(this.client)
-	  WpNotificationRepository.onServiceAssigned(this.serviceAssigned).catch(e => Sentry.captureException(e))
-	  WpNotificationRepository.onDriverArrived(this.driverArrived).catch(e => Sentry.captureException(e))
-	  WpNotificationRepository.onNewService(this.onNewService).catch(e => Sentry.captureException(e))
-    if (this.socket) this.socket.emit(Events.READY)
+	  WpNotificationRepository.onServiceAssigned(this.wpClient.id, this.serviceAssigned).catch(e => Sentry.captureException(e))
+	  WpNotificationRepository.onDriverArrived(this.wpClient.id, this.driverArrived).catch(e => Sentry.captureException(e))
+	  WpNotificationRepository.onNewService(this.wpClient.id, this.onNewService).catch(e => Sentry.captureException(e))
+    if (this.socket) this.socket.to(this.wpClient.id).emit(Events.READY)
     console.table(this.client.pupBrowser?._targets)
   }
 
   onQR = (qr: string): void => {
-    if (this.socket) this.socket.emit(Events.QR_RECEIVED, qr)
-    console.log('sending qr code..', qr)
+    if (this.socket) this.socket.to(this.wpClient.id).emit(Events.QR_RECEIVED, qr)
+    console.log('sending qr code..', this.wpClient.alias, qr)
   }
   
   onAuth = (): void => {
-		console.log('authentication successfully!')
+		console.log('authentication successfully!', this.wpClient.alias)
   }
 	
 	onMessageReceived = (msg: Message): void => {
 		if (this.isProcessableMsg(msg)) SessionRepository.addChat(msg).catch((e) => {
-			console.log('msg message', msg.type, msg.from)
 			console.warn('error saving message', e.message)
 		})
 	}
 
 	isProcessableMsg(msg: Message): boolean {
-		return (msg.type === MessageTypes.TEXT && !msg.from.includes('-'))
+		return (msg.type === MessageTypes.TEXT && !msg.from.includes('-')) || msg.type === MessageTypes.LOCATION
 	}
-  
+
   onDisconnected = async (reason: string | WAState): Promise<void> => {
-    console.log('Client disconnected ', reason)
-		await SettingsRepository.enableWpNotifications(false)
-    if (this.socket) this.socket.emit(Events.DISCONNECTED, reason)
+    console.log('Client disconnected ', this.wpClient.alias, reason)
+		await SettingsRepository.enableWpNotifications(this.wpClient.id, false)
+    if (this.socket) this.socket.to(this.wpClient.id).emit(Events.DISCONNECTED, reason)
     if (reason === EmitEvents.NAVIGATION) await this.client.destroy().catch(e => {
-			console.log('destroy ', e.message)
+			console.log('destroy ', this.wpClient.alias, e.message)
 			this.socket?.emit(EmitEvents.FAILURE, e.message)
 			Sentry.captureException(e)
 			exit(1)
@@ -114,8 +130,8 @@ export default class WhatsAppClient {
   }
   
   onAuthFailure = (message: string): void => {
-    if (this.socket) this.socket.emit(Events.AUTHENTICATION_FAILURE, message)
-    console.log(Events.AUTHENTICATION_FAILURE, message)
+    if (this.socket) this.socket.to(this.wpClient.id).emit(Events.AUTHENTICATION_FAILURE, message)
+    console.log(Events.AUTHENTICATION_FAILURE, this.wpClient.alias, message)
   }
 	
 	onLoadingScreen = (percent: string, message: string): void => {
@@ -123,41 +139,42 @@ export default class WhatsAppClient {
 			percent: percent,
 			message: message
 		}
-		if (this.socket) this.socket.emit(Events.LOADING_SCREEN, loading)
-		console.log(Events.LOADING_SCREEN, percent, message)
+		if (this.socket) this.socket.to(this.wpClient.id).emit(Events.LOADING_SCREEN, loading)
+		console.log(Events.LOADING_SCREEN, this.wpClient.alias, percent, message)
 	}
 	
   onStateChanged = (waState: WAState): void => {
-    if (this.socket) this.socket.emit(EmitEvents.GET_STATE, waState)
-    console.log(Events.STATE_CHANGED, waState)
+    if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, waState)
+    console.log(Events.STATE_CHANGED, this.wpClient.alias, waState)
   }
   
   init = async (web = true): Promise<void> => {
-    console.log('initializing whatsapp client...')
-		if (this.socket) this.socket.emit(EmitEvents.GET_STATE, WAState.OPENING)
-		if (web && !this.client.pupPage?.isClosed()) await this.client.destroy().catch(e => console.log(e))
+		this.starting = true
+    console.log('initializing whatsapp client...', this.wpClient.alias)
+		if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, WAState.OPENING)
+		if (web && !this.client.pupPage?.isClosed()) await this.client.destroy().catch(e => console.log(e, this.wpClient.alias))
     return this.client.initialize()
   }
   
   getState = (): void => {
     this.client.getState().then(state => {
-      if (this.socket) this.socket.emit(EmitEvents.GET_STATE, state)
+      if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, state)
     }).catch(e => {
-      console.log(EmitEvents.GET_STATE, e.message)
-      if (this.socket) this.socket.emit(EmitEvents.GET_STATE, WAState.UNPAIRED)
+      console.log(EmitEvents.GET_STATE, this.wpClient.alias, e.message)
+      if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, WAState.UNPAIRED)
     })
   }
   
   serviceAssigned = async (snapshot: DataSnapshot): Promise<void> => {
     const notification: WpNotificationType = snapshot.val()
-    if (notification.driver_id != null) {
+    if (notification.driver_id != null && notification.wp_client_id == this.wpClient.id) {
       const driver = this.store.findDriverById(notification.driver_id)
       await this.client.sendMessage(notification.client_id, Messages.serviceAssigned(driver.vehicle)).then(() => {
 				WpNotificationRepository.deleteNotification('assigned', snapshot.key?? '')
 			}).catch(async e => {
-				console.log('serviceAssigned', e)
+				console.log('serviceAssigned', this.wpClient.alias, e)
 				Sentry.captureException(e)
-				if (this.socket) this.socket.emit(EmitEvents.GET_STATE, WAState.OPENING)
+				if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, WAState.OPENING)
 				exit(1)
 			})
     } else {
@@ -170,9 +187,9 @@ export default class WhatsAppClient {
     await this.client.sendMessage(notification.client_id, Messages.DRIVER_ARRIVED).then(() => {
 			WpNotificationRepository.deleteNotification('arrived', snapshot.key?? '')
 		}).catch(async e => {
-			console.log('driverArrived', e)
+			console.log('driverArrived', this.wpClient.alias, e)
 			Sentry.captureException(e)
-			if (this.socket) this.socket.emit(EmitEvents.GET_STATE, WAState.OPENING)
+			if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, WAState.OPENING)
 			exit(1)
 		})
   }
@@ -182,10 +199,9 @@ export default class WhatsAppClient {
     await this.client.sendMessage(notification.client_id, Messages.CANCELED).then(() => {
 			WpNotificationRepository.deleteNotification('canceled', snapshot.key?? '')
 		}).catch(async e => {
-			console.log('serviceCanceled', e)
+			console.log('serviceCanceled', this.wpClient.alias, e)
 			Sentry.captureException(e)
-			await SettingsRepository.enableWpNotifications(false)
-			if (this.socket) this.socket.emit(EmitEvents.GET_STATE, WAState.OPENING)
+			if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, WAState.OPENING)
 			exit(1)
 		})
   }
@@ -195,10 +211,9 @@ export default class WhatsAppClient {
     await this.client.sendMessage(notification.client_id, Messages.SERVICE_COMPLETED).then(() => {
 			WpNotificationRepository.deleteNotification('terminated', snapshot.key?? '')
 		}).catch(async e => {
-			console.log('serviceTerminated', e)
+			console.log('serviceTerminated', this.wpClient.alias, e)
 			Sentry.captureException(e)
-			await SettingsRepository.enableWpNotifications(false)
-			if (this.socket) this.socket.emit(EmitEvents.GET_STATE, WAState.OPENING)
+			if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, WAState.OPENING)
 			exit(1)
 		})
   }
@@ -209,9 +224,9 @@ export default class WhatsAppClient {
 		await this.client.sendMessage(notification.client_id, Messages.NEW_SERVICE).then(() => {
 			WpNotificationRepository.deleteNotification('new', snapshot.key?? '')
 		}).catch(async e => {
-			console.log('onNewService', e)
+			console.log('onNewService', this.wpClient.alias, e)
 			Sentry.captureException(e)
-			if (this.socket) this.socket.emit(EmitEvents.GET_STATE, WAState.OPENING)
+			if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, WAState.OPENING)
 			exit(1)
 		})
 	}
@@ -222,9 +237,9 @@ export default class WhatsAppClient {
 			ServiceRepository.findServiceStatusById(serviceId).then((status) => {
 				if (status === Service.STATUS_PENDING) {
 					this.client.sendMessage(clientId, ASK_FOR_CANCEL).catch(e => {
-						console.log('cancelTimeout', e)
+						console.log('cancelTimeout', this.wpClient.alias, e)
 						Sentry.captureException(e)
-						if (this.socket) this.socket.emit(EmitEvents.GET_STATE, WAState.OPENING)
+						if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.GET_STATE, WAState.OPENING)
 						exit(1)
 					})
 				}
@@ -233,15 +248,56 @@ export default class WhatsAppClient {
 	}
   
   logout = (): void => {
-    this.client.logout()
+    this.client.destroy()
       .then(() => {
-        if (this.socket) this.socket.emit('destroy')
+        if (this.socket) this.socket.to(this.wpClient.id).emit('destroy')
       })
       .catch(e => {
-			console.log('logout: ', e)
-			if (this.socket) this.socket.emit(EmitEvents.FAILURE, e.message)
+			console.log('logout: ', this.wpClient.alias, e)
+			if (this.socket) this.socket.to(this.wpClient.id).emit(EmitEvents.FAILURE, e.message)
 			Sentry.captureException(e)
 			exit(1)
 		})
   }
+
+	setWpClient(client: WpClient): void {
+		this.wpClient.wpNotifications = client.wpNotifications
+		this.wpClient.chatBot = client.chatBot
+	}
+
+	serviceChanged = async (snapshot: DataSnapshot): Promise<void> => {
+		if (!this.wpClient.chatBot) return
+		const service = new Service()
+		Object.assign(service, snapshot.val() as ServiceInterface)
+		const session = new Session(service.client_id)
+		session.service_id = service.id
+		session.status = Session.STATUS_REQUESTING_SERVICE
+		let sessionDB = await SessionRepository.findSessionByChatId(service.client_id)
+		if (!sessionDB) {
+			sessionDB = await SessionRepository.create(session)
+		}
+		Object.assign(session, sessionDB)
+		switch (service.status) {
+			case Service.STATUS_IN_PROGRESS:
+				const driver = this.store.findDriverById(service.driver_id!!)
+				if (service.metadata && service.metadata.arrived_at > 0 && service.metadata.start_trip_at == null) {
+					await this.client.sendMessage(service.client_id, Messages.DRIVER_ARRIVED)
+				} else if (!service.metadata) {
+					await session.setStatus(Session.STATUS_SERVICE_IN_PROGRESS)
+					await this.client.sendMessage(service.client_id,
+						Messages.serviceAssigned(driver.vehicle))
+				}
+				break
+			case Service.STATUS_TERMINATED:
+				await session.setStatus(Session.STATUS_COMPLETED)
+				await this.client.sendMessage(service.client_id, Messages.SERVICE_COMPLETED)
+				break
+			case Service.STATUS_CANCELED:
+				await session.setStatus(Session.STATUS_COMPLETED)
+				await this.client.sendMessage(service.client_id, Messages.CANCELED)
+				break
+			default:
+				console.log('new service', service.id)
+		}
+	}
 }
